refactor: SigningUtil static methods into functions
