+++ packages/client/src/Authentication.ts
@@ -9,7 +9,7 @@ import { XOR } from './types'
 import { pLimitFn } from './utils/promises'
 import pMemoize from 'p-memoize'
 import { wait } from '@streamr/utils'
-import SigningUtil from './utils/SigningUtil'
+import { sign } from './utils/signingUtils'
 
 export type ProviderConfig = ExternalProvider
 
@@ -47,7 +47,7 @@ export const createAuthentication = (authConfig: AuthConfig, ethereumConfig: Eth
         return {
             isAuthenticated: () => true,
             getAddress: async () => address,
-            createMessagePayloadSignature: async (payload: string) => SigningUtil.sign(payload, key),
+            createMessagePayloadSignature: async (payload: string) => sign(payload, key),
             getStreamRegistryChainSigner: async () => new Wallet(key, getStreamRegistryChainProvider(ethereumConfig))
         }
     } else if (authConfig.ethereum !== undefined) {
